+++ HISTORY.md
@@ -9,6 +9,8 @@
 - Implemented support for fractions, powered by the library `fraction.js`.
 - Implemented matrix LU decomposition with partial pivoting and a LU based 
   linear equations solver (functions `lup` and `lusolve`). Thanks @rjbaucells.
+- Implemented a new configuration option `predictable`, which can be set to
+  true in order to ensure predictable function output types.
 - Large internal refactoring, allowing to create custom bundles of math.js.
 - Function `typeof` no longer returns lower case names, but now returns lower
   case names for primitives (like `number`, `boolean`, `string`), and